moved testutil into syslutil
